ai/pyproject.toml
[tool.poetry]
name = "botchat-ai"
version = "0.1.0"
description = "A FastAPI-based PDF RAG system with Qdrant vector database and LLM integration"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

# Core dependencies
fastapi = "^0.115.14"
uvicorn = "^0.34.3"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"

# Document processing
pypdf = "^5.6.1"
pypdf2 = "^3.0.1"
pymupdf = "^1.26.1"
docx2txt = "^0.8"
python-pptx = "^0.6.23"
python-magic = "^0.4.27"
tabulate = "^0.9.0"

# LangChain ecosystem
langchain = "^0.0.353"  # Using a stable version
langchain-community = "^0.0.16"  # Compatible version
langchain-core = "^0.1.8"  # Compatible version

# Vector database
qdrant-client = "1.8.0"  # Version known to work well with the above

# Embeddings
sentence-transformers = "^4.1.0"

# Database
pymongo = "^4.13.2"
motor = "^3.7.1"
redis = "4.6.0"

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"

# Other utilities
requests = "^2.32.4"
google-generativeai = "^0.8.5"
einops = "^0.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.2.0"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
